refactor(multichain): move stateChange event subscription into BalancesController
